LRQA-30742 Move to amazon package
