+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/vm/provisioner/amazon/BaseAuroraAwsVmProvisioner.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.jenkins.results.parser.vm.provisioner;
+package com.liferay.jenkins.results.parser.vm.provisioner.amazon;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
@@ -32,6 +32,7 @@ import com.amazonaws.services.rds.model.DescribeDBInstancesResult;
 import com.amazonaws.services.rds.model.Endpoint;
 
 import com.liferay.jenkins.results.parser.JenkinsResultsParserUtil;
+import com.liferay.jenkins.results.parser.vm.provisioner.VMProvisioner;
 
 import java.util.List;
